% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdf_adv.R
\name{cdf_adv}
\alias{cdf_adv}
\title{CDV and ADV}
\usage{
cdf_adv(
  data,
  subsample,
  important = c(0:5000),
  delta = 1,
  theta = 1,
  epsilon = 0,
  gamma = 0.1,
  reference = TRUE,
  plotting = TRUE
)
}
\arguments{
\item{data}{original sample data}

\item{subsample}{subsample}

\item{important}{length classes to be retained in the sub-sampling}

\item{delta}{helps to reveal robust modes and antimodes of LFD}

\item{theta}{keeps distinguished differences between adjacent robust modes and antimodes}

\item{epsilon}{chooses an interval, where locations of robust modes and antimodes for larger specimens may vary}

\item{gamma}{indicates a minimally required number per length class in a reference subsample}

\item{reference}{TRUE or FALSE}

\item{plotting}{TRUE or FALSE}
}
\value{
distance between the cumulative of the original sample and the sub-samples
}
\description{
CDV and ADV
}
\examples{
cdf_adv(filtered.sample, reference.subsample)
}
