% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataRDBEStoRDB.R
\name{dataRDBEStoRDB}
\alias{dataRDBEStoRDB}
\title{dataRDBEStoRDB}
\usage{
dataRDBEStoRDB(x, selected_upperHierarchy = 1)
}
\arguments{
\item{x}{zip file containi RDBES input files}

\item{selected_upperHierarchy}{upper hierarchy}
}
\value{
list(HL = HL, HH = HH, SL = SL, CA = CA) RDB tables
}
\description{
dataRDBEStoRDB
}
\examples{
dataRDBEStoRDB(list.rdbes, 1)

}
